Refactor CustomFields type in its own module
